==> 123_steps/code/00_run_all.R <==
getwd()
source("./code/02_Note7.R")
source("./code/03_Note7.R")

source("./code/02_Hoverboard.R")
source("./code/03_Hoverboard.R")

source("./code/02_Toyota.R")
source("./code/03_Toyota.R")

source("./code/02_BlueBell.R")
source("./code/03_BlueBell.R")

source("./code/02_Mars.R")
source("./code/03_Mars.R")

source("./code/02_Whirpool.R")
source("./code/03_Whirpool.R")
